Introduce TailwindCss 😬
Giving tailwind a go. Usually MUI is my go to, but I have never really
been a fan of CSS-in-JS, and also it has some limitations when it comes
to using with remix (Mainly can't cache css as it does it is injected
into html and required double rendering to SSR properly).

Using the CubeCSS methodology with TailwindCSS will hopefully help as I
am not a fan of overusing utilities in html.

Let's see how this goes...